requirements.txt
# Web Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0

# HTTP Requests
requests==2.31.0

# Web Scraping
beautifulsoup4==4.12.3
newspaper3k==0.2.8
lxml==5.1.0

# OpenAI
openai==1.12.0

# Environment Variables
python-dotenv==1.0.1

# Utilities
python-multipart==0.0.9